In computer class, the first assignment was to write a program to print the first 100 Fibonacci numbers. Instead, I wrote a program that would steal passwords of students. My teacher gave me an A

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -5,15 +5,15 @@
  */
 void print_fibonacci(void)
 {
-	unsigned long long int first = 1, second = 2, next;
+	unsigned int first = 1, second = 2, next;
 	int i;
 
 	for (i = 0; i < 50; i++)
 	{
 		if (i < 49)
-			printf("%llu, ", first);
+			printf("%u, ", first);
 		else
-			printf("%llu\n", first);
+			printf("%u\n", first);
 
 		next = first + second;
 		first = second;
